+++ simulation/entities/Entity_Thwomp.cpp
@@ -57,7 +57,7 @@ bool Entity_Thwomp::CollideVsCircle_Logical(Simulator* sim, Ninja* ninja, collis
 void Entity_Thwomp::Think(Simulator* sim) {
     if (CUR_STATE == 0) {
         Grid_Edges* edgeGrid = sim->edgeGrid;
-        for (int i = 0; i < sim->playerList.size(); i++) {
+        for (size_t i = 0; i < sim->playerList.size(); i++) {
             Ninja* ninja = sim->playerList[i];
             if (!ninja->IsDead()) {
                 vec2 ninjaPos = ninja->GetPos();
